requirements.txt
javalang==0.13.0
six==1.16.0

python-dotenv~=1.0.1
requests~=2.32.3